Parallelize downloads

+++ generate_librimix.sh
@@ -1,41 +1,68 @@
 #!/bin/bash
-set -e  # Exit on error
+set -eu  # Exit on error
 
 storage_dir=$1
-
-echo "Download LibiSpeech/dev-clean into $storage_dir"
-# If downloading stalls for more than 20s, relaunch from previous state.
-wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/dev-clean.tar.gz -P $storage_dir
-tar -xzf $storage_dir/dev-clean.tar.gz -C $storage_dir
-rm -rf $storage_dir/dev-clean.tar.gz
-
-echo "Download LibiSpeech/test-clean into $storage_dir"
-# If downloading stalls for more than 20s, relaunch from previous state.
-wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/test-clean.tar.gz -P $storage_dir
-tar -xzf $storage_dir/test-clean.tar.gz -C $storage_dir
-rm -rf $storage_dir/test-clean.tar.gz
-
-echo "Download LibiSpeech/train-clean-100 into $storage_dir"
-# If downloading stalls for more than 20s, relaunch from previous state.
-wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/train-clean-100.tar.gz -P $storage_dir
-tar -xzf $storage_dir/train-clean-100.tar.gz -C $storage_dir
-rm -rf $storage_dir/train-clean-100.tar.gz
-
-echo "Download LibiSpeech/train-clean-360 into $storage_dir"
-# If downloading stalls for more than 20s, relaunch from previous state.
-wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/train-clean-360.tar.gz -P $storage_dir
-tar -xzf $storage_dir/train-clean-360.tar.gz -C $storage_dir
-rm -rf $storage_dir/train-clean-360.tar.gz
-
-echo "Download wham_noise into $storage_dir"
-# If downloading stalls for more than 20s, relaunch from previous state.
-wget -c --tries=0 --read-timeout=20 https://storage.googleapis.com/whisper-public/wham_noise.zip -P $storage_dir
-unzip -qn $storage_dir/wham_noise.zip -d $storage_dir
-rm -rf $storage_dir/wham_noise.zip
-
 librispeech_dir=$storage_dir/LibriSpeech
 wham_dir=$storage_dir/wham_noise
 librimix_outdir=$storage_dir/
+
+function LibriSpeech_dev_clean() {
+	if ! test -e $librispeech_dir/dev-clean; then
+		echo "Download LibriSpeech/dev-clean into $storage_dir"
+		# If downloading stalls for more than 20s, relaunch from previous state.
+		wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/dev-clean.tar.gz -P $storage_dir
+		tar -xzf $storage_dir/dev-clean.tar.gz -C $storage_dir
+		rm -rf $storage_dir/dev-clean.tar.gz
+	fi
+}
+
+function LibriSpeech_test_clean() {
+	if ! test -e $librispeech_dir/test-clean; then
+		echo "Download LibriSpeech/test-clean into $storage_dir"
+		# If downloading stalls for more than 20s, relaunch from previous state.
+		wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/test-clean.tar.gz -P $storage_dir
+		tar -xzf $storage_dir/test-clean.tar.gz -C $storage_dir
+		rm -rf $storage_dir/test-clean.tar.gz
+	fi
+}
+
+function LibriSpeech_clean100() {
+	if ! test -e $librispeech_dir/train-clean-100; then
+		echo "Download LibriSpeech/train-clean-100 into $storage_dir"
+		# If downloading stalls for more than 20s, relaunch from previous state.
+		wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/train-clean-100.tar.gz -P $storage_dir
+		tar -xzf $storage_dir/train-clean-100.tar.gz -C $storage_dir
+		rm -rf $storage_dir/train-clean-100.tar.gz
+	fi
+}
+
+function LibriSpeech_clean360() {
+	if ! test -e $librispeech_dir/train-clean-360; then
+		echo "Download LibriSpeech/train-clean-360 into $storage_dir"
+		# If downloading stalls for more than 20s, relaunch from previous state.
+		wget -c --tries=0 --read-timeout=20 http://www.openslr.org/resources/12/train-clean-360.tar.gz -P $storage_dir
+		tar -xzf $storage_dir/train-clean-360.tar.gz -C $storage_dir
+		rm -rf $storage_dir/train-clean-360.tar.gz
+	fi
+}
+
+function wham() {
+	if ! test -e $wham_dir; then
+		echo "Download wham_noise into $storage_dir"
+		# If downloading stalls for more than 20s, relaunch from previous state.
+		wget -c --tries=0 --read-timeout=20 https://storage.googleapis.com/whisper-public/wham_noise.zip -P $storage_dir
+		unzip -qn $storage_dir/wham_noise.zip -d $storage_dir
+		rm -rf $storage_dir/wham_noise.zip
+	fi
+}
+
+LibriSpeech_dev_clean &
+LibriSpeech_test_clean &
+LibriSpeech_clean100 &
+LibriSpeech_clean360 &
+wham &
+
+wait
 
 
 python scripts/augment_train_noise.py --wham_dir $wham_dir
